general file restructure for clarity
