+++ app/(root)/page.tsx
@@ -1,10 +1,10 @@
-import Hero from '@/components/hero/Hero';
-import Skills from '@/components/skills/Skills';
-import Services from '@/components/service/Services';
-import Work from '@/components/work/Work';
-import Project from '@/components/projects/Project';
-import Testimonials from '@/components/testimonials/Testimonials';
-import CallToAction from '@/components/action/CallToAction';
+import Hero from '@/components/homepage/hero/Hero';
+import Skills from '@/components/homepage/skills/Skills';
+import Services from '@/components/homepage/service/Services';
+import Work from '@/components/homepage/work/Work';
+import Project from '@/components/homepage/projects/Project';
+import Testimonials from '@/components/homepage/testimonials/Testimonials';
+import CallToAction from '@/components/shared-components/action/CallToAction';
 
 const Home = () => {
   return (
