+++ components/case-studies/index.ts
@@ -1,4 +1,4 @@
 export { default as CaseStudiesHeading } from './CaseStudiesHeadingContent';
 export { default as CaseStudiesCard } from './CaseStudiesCard';
 export { getCaseStudies } from '@/sanity/sanity.query';
-export { default as CallToAction } from '../action/CallToAction';
+export { default as CallToAction } from '../shared-components/action/CallToAction';
