+++ components/homepage/hero/ClipBoardText.tsx
