+++ components/homepage/hero/HeroHeadingContent.tsx
