+++ components/homepage/projects/ProjectCardContent.tsx
