+++ components/project-details/caseStudiesPreview/Preview.tsx
@@ -1,9 +1,9 @@
 import WrapperProjectDetails from '@/HOC/WrapperProjectDetails';
-import { CaseStudiesType } from '@/types';
+import { PreviewType } from '@/types';
 import SectionHeading from '../../reusable/SectionHeading';
-import StudyCard from './StudyCard';
+import StudyCard from './PreviewCard';
 
-const CaseStudies = ({ otherCaseStudies }: CaseStudiesType) => {
+const Preview = ({ otherCaseStudies }: PreviewType) => {
   return (
     <WrapperProjectDetails additionalStyles="bg-white-800 dark:bg-black-300 px-[1.5rem] py-[2.25rem] md:py-[4.5rem] md:px-[4.5rem]">
       <SectionHeading
@@ -11,7 +11,7 @@ const CaseStudies = ({ otherCaseStudies }: CaseStudiesType) => {
         mainHeading="Other Case Studies"
         additionalStylesMainHeading="mb-[2.25rem] md:pb-[2.5rem]"
       />
-      <div className="flex flex-col items-center justify-between sm:flex-row">
+      <div className="flex flex-col items-center justify-between dark:bg-black-200 sm:flex-row">
         {otherCaseStudies.map((study, index) => (
           <StudyCard key={index} {...study} />
         ))}
@@ -20,4 +20,4 @@ const CaseStudies = ({ otherCaseStudies }: CaseStudiesType) => {
   );
 };
 
-export default CaseStudies;
+export default Preview;
