+++ components/project-details/index.ts
@@ -1,6 +1,6 @@
 export { default as Header } from './details/Header';
 export { default as DetailsLinks } from './details/DetailsLinks';
-export { default as TechStack } from './details/TechStack';
+export { default as TechStack } from './tech-stack/TechStack';
 export { default as Timeline } from './details/Timeline';
 export { default as ProblemStatement } from './details/ProblemStatement';
 export { default as Process } from './process/Process';
@@ -8,6 +8,6 @@ export { default as DetailsImage } from './details/DetailsImage';
 export { default as Description } from './details/Description';
 export { default as FigmaDesign } from './figma/FigmaDesign';
 export { default as ChallengesSection } from './challenges/ChallengesSection';
-export { default as CallToAction } from '../action/CallToAction';
-export { default as CaseStudies } from './caseStudies/CaseStudies';
+export { default as CallToAction } from '../shared-components/action/CallToAction';
+export { default as CaseStudies } from './caseStudiesPreview/Preview';
 export { default as TopContent } from './details/TopContent';
