+++ components/project-details/tech-stack/TechSkillIcon.tsx
