+++ components/project-details/tech-stack/TechStack.tsx
