+++ components/shared-components/action/CallToActionButton.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 
-import CustomButton from '../reusable/CustomButton';
+import CustomButton from '../../reusable/CustomButton';
 
 const CallToActionButton = () => {
   return (
